@@ -13,7 +13,7 @@ project_urls =
 classifiers =
     Development Status :: 5 - Production/Stable
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.12
     Programming Language :: Python :: Implementation :: CPython
     License :: OSI Approved :: BSD License
     Operating System :: OS Independent
@@ -23,63 +23,65 @@ package_dir =
     = src
 packages = find:
 include_package_data = True
-python_requires = >=3.10
+python_requires = >=3.12
 install_requires =
-    aiodns ==3.1.1
-    aiohttp ==3.9.3
-    async-timeout ==4.0.3
-    certifi ==2024.2.2
-    cffi ==1.16.0
-    charset-normalizer ==3.3.2
-    discord.py ==2.3.2
-    cryptography ==42.0.5
-    gevent ==24.2.1
-    greenlet ==3.0.3
-    idna ==3.6
-    oauthlib ==3.2.2
-    redis ==4.6.0
-    requests ==2.31.0
-    requests-oauthlib ==1.3.1
-    psycopg[c] ==3.1.18
-    psycopg-pool ==3.2.1
-    pydantic ==2.6.3
-    pydantic-core ==2.16.3
-    pydantic-settings ==2.2.1
-    PyYAML ==6.0.1
-    typing_extensions ==4.10.0
-    urllib3 ==2.2.1
-    zope.event ==5.0
-    zope.interface ==6.2
+    aiodns
+    aiohttp
+    async-timeout
+    certifi
+    cffi
+    charset-normalizer
+    discord.py
+    cryptography
+    gevent
+    greenlet
+    idna
+    oauthlib
+    redis
+    requests
+    requests-oauthlib
+    psycopg[c]
+    psycopg-pool
+    pydantic
+    pydantic-core
+    pydantic-settings
+    PyYAML
+    typing_extensions
+    urllib3
+    zope.event
+    zope.interface
 
 [options.extras_require]
-tests =
-    bandit[toml] ==1.7.7
-    coverage[toml] ==7.4.3
-    dpytest ==0.7.0
-    flake8 ==7.0.0
-    flake8-bugbear == 24.2.6
-    flaky ==3.7.0
-    mypy ==1.8.0
-    pre-commit ==3.6.2
-    pytest ==8.0.2
-    pytest-asyncio ==0.23.5
-    pytest-cov ==4.1.0
-    pytest-gevent ==1.1.0
-    pytest-postgresql ==5.1.0
-    pytest-redis ==3.0.2
-    time-machine ==2.13.0
-    types-gevent ==24.2.0.20240301
-    types-greenlet ==3.0.0.20240106
-    types-oauthlib ==3.2.0.20240217
-    types-psutil ==5.9.5.20240205
-    types-pyOpenSSL ==24.0.0.20240228
-    types-pyYAML ==6.0.12.12
-    types-redis ==4.6.0.20240218
-    types-requests ==2.31.0.20240218
-    types-requests-oauthlib ==1.3.0.20240106
-    types-uWSGI ==2.0.0.20240106
+dev =
+    bandit[toml]
+    coverage[toml]
+    dpytest
+    flake8
+    flake8-bugbear
+    flaky
+    mypy
+    pre-commit
+    pre-commit-uv
+    pytest
+    pytest-asyncio
+    pytest-cov
+    pytest-gevent
+    pytest-postgresql
+    pytest-redis
+    time-machine
+    types-gevent
+    types-greenlet
+    types-oauthlib
+    types-psutil
+    types-pyOpenSSL
+    types-pyYAML
+    types-redis
+    types-requests
+    types-requests-oauthlib
+    types-uWSGI
+    uv
 uwsgi =
-    uwsgi ==2.0.24
+    uwsgi
 
 [options.package_data]
 turbo_sticks =
